--- a/app/models/flavor.rb
+++ b/app/models/flavor.rb
@@ -45,7 +45,7 @@ class Flavor < ApplicationRecord
 
   def self.tenant_joins_clause(scope)
     scope.includes(:cloud_tenants => "source_tenant", :ext_management_system => {})
-         .references(:cloud_tenants => "source_tenant", :ext_management_system => {})
+         .references(:cloud_tenants, :tenants, :ext_management_system)
   end
 
   def self.create_flavor_queue(userid, ext_management_system, options = {})