@kbrock 2019-06-13
update: rebased out extracted PR
updated top comments