@jrafanie 2019-06-14
@kbrock I'm having a hard time reviewing this.    Is there a bz for this?  Where is this most a beneficial?  I recall report_data being an endpoint hit by possibly both the UI and api.  Can you demonstrate a common report before/after?  Does this break support for anything?

I don't know what the numbers mean in terms of how much faster this change is.