@kbrock 2019-06-19
wip: need to fix deep_merge and further investigate inner_sql construction of joins when not all fields are in the requires